backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy>=2.0.38
psycopg[binary]==3.2.10
alembic==1.14.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
argon2-cffi==23.1.0
pydantic-settings==2.7.1
email-validator==2.2.0
python-multipart==0.0.20
pytest==8.3.5
httpx==0.28.1
